requirements.txt
easydict==1.13
addict==2.4.0
einops==0.8.1
tokenizers==0.20.3
transformers==4.46.3
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
datasets>=2.0.0
requests>=2.28.0
tqdm>=4.60.0
numpy>=1.21.0
matplotlib>=3.5.0
datasets==4.3.0
jiwer>=4.0.0
